pyproject.toml
[project]
name = "mcp-nano-banana"
version = "0.3.3"
requires-python = "==3.10"
description = "Um servidor MCP que gera imagens usando a Google Gemini API."
readme = "README.md"
dependencies = [
    "mcp[cli]==1.13",
    "google-generativeai==0.5.4",
    "python-dotenv==1.0.1",
    "pillow==11.3.0",
]

[project.urls]
Homepage = "https://github.com/GuilhermeAumo/mcp-nano-banana"
Repository = "https://github.com/GuilhermeAumo/mcp-nano-banana"

[project.scripts]
mcp-nano-banana = "mcp_nano_banana.main:main"

[tool.setuptools.packages.find]
where = ["src"]
